backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
python-multipart
pyjwt
passlib[bcrypt]
pillow
aiofiles